backend/pyproject.toml
[project]
name = "notes-backend"
version = "0.1.0"
description = "Note-sharing application with expiration and password protection"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.134",
    "uvicorn[standard]>=0.41",
    "google-cloud-firestore>=2.23",
    "pydantic>=2.12",
    "pydantic-settings>=2.13",
    "email-validator>=2.2",
    "argon2-cffi>=25.1",
    "pyjwt>=2.11",
    "httpx>=0.28",
    "itsdangerous>=2.2",
    "slowapi>=0.1",
    "aiosqlite>=0.22.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3",
    "pytest-cov>=6.0",
    "testcontainers[google]>=4.8",
    "httpx>=0.28",
    "ruff>=0.15",
    "pyright>=1.1",
    "vulture>=2.14",
    "pre-commit>=4.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py314"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "ASYNC", "ANN", "S", "C4", "PT"]
ignore = ["ANN401"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106", "S110", "ANN"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-report=html"
markers = [
    "unit: fast deterministic unit tests",
    "integration: tests requiring external services (e.g. testcontainers)",
]
filterwarnings = [
    "ignore::DeprecationWarning:google.*",
    "ignore::DeprecationWarning:pkg_resources.*",
    "ignore::DeprecationWarning:pydantic.*",
]

[tool.coverage.run]
branch = true
source = ["app"]
omit = ["app/jobs/*"]

[tool.coverage.report]
show_missing = true
fail_under = 0

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "standard"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnnecessaryTypeIgnoreComment = false
reportUnusedImport = false
venvPath = "."
venv = ".venv"

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101", "B105", "B110"]
